fix(ui): stabilize select dropdown sizing and panel positioning
Extract a shared SearchableSelect hook that sizes panels to the widest
option, add combobox min-width for short values, keep full width in forms
while shrink-wrapping inline toolbar filters, and document edge cases in Storybook.
